back/requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
APScheduler==3.10.4
asyncpg==0.29.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
fastapi==0.114.1
filelock==3.16.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.0
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
jiter==0.5.0
jsonpath-python==1.0.6
mistralai==1.0.3
mypy-extensions==1.0.0
openai==1.44.1
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
replicate==0.32.1
requests==2.32.3
rsa==4.9
sentry-sdk==2.14.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
tokenizers==0.20.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
